change all Int64 to Int
Which is Int32 on 32 bit systems. Not sure if we will support 32 bit. But let's use Int for brevity.

+++ src/io.jl
@@ -156,7 +156,7 @@ end
 struct NCReader{T}
     dataset::NCDataset
     cyclic_dataset::NCDataset
-    cyclic_times::Vector{Tuple{Int64,Int64}}
+    cyclic_times::Vector{Tuple{Int,Int}}
     forcing_parameters::Dict{Symbol,String}
     cyclic_parameters::Dict{Symbol,String}
     buffer::Matrix{T}
